+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   This change is made to be consistent with Linux kernel and wireguard-go.
 - Continue with IPv4-only UDP transport on Linux when IPv6 sockets are unavailable.
 
+### Removed
+- Remove `AsFd` implementation for `UdpSocket`.
+
 
 ## [0.7.2] - 2026-06-25
 ### Added